backend/requirements.txt
openai
feedparser
requests
python-dotenv
supabase
python-telegram-bot
beautifulsoup4
lxml
